+++ packages/transaction-controller/src/index.ts
@@ -107,10 +107,6 @@ export {
   UserFeeLevel,
   WalletDevice,
 } from './types';
-export {
-  DISPLAYED_TRANSACTION_HISTORY_PATHS,
-  MAX_TRANSACTION_HISTORY_LENGTH,
-} from './utils/history';
 export { determineTransactionType } from './utils/transaction-type';
 export { mergeGasFeeEstimates } from './utils/gas-flow';
 export {